divided in column elements
